fix(cloud): narrow snowflake_id to 53-bit for JS Number safety and fix store_info query type
snowflake_id was 63-bit (41+22) which exceeds Number.MAX_SAFE_INTEGER
causing precision loss in Console JS. Now 53-bit (41+12) — still 4096
values/ms which is collision-free at POS scale.

Also cast store_info SELECT `1 AS id` to `1::BIGINT` to match the i64
column type expected by sqlx query_as.

+++ crab-cloud/src/db/store/store_info.rs
@@ -111,7 +111,7 @@ pub async fn get_store_info(
 ) -> Result<Option<StoreInfo>, BoxError> {
     let row: Option<StoreInfo> = sqlx::query_as(
         r#"
-        SELECT 1 AS id, name, address, nif, logo_url, phone, email, website,
+        SELECT 1::BIGINT AS id, name, address, nif, logo_url, phone, email, website,
                business_day_cutoff, created_at, updated_at
         FROM store_info
         WHERE edge_server_id = $1
